# internal/keyspan: preallocate merging iter levels, items