Take care of a pending TODO, and avoid allocations when the range-key iterator
stack has any child iterators. This has a minor effect on the MVCCGet
benchmarks that use a batch, which requires the batch to be included within the
range key merging iter.

```
name                                                    old time/op    new time/op    delta
MVCCGet_Pebble/batch=false/versions=1/valueSize=8-10      4.03µs ± 1%    4.06µs ± 1%  +0.61%  (p=0.026 n=9+10)
MVCCGet_Pebble/batch=false/versions=10/valueSize=8-10     4.90µs ± 1%    4.86µs ± 4%    ~     (p=0.210 n=10+10)
MVCCGet_Pebble/batch=false/versions=100/valueSize=8-10    9.91µs ± 2%    9.97µs ± 2%    ~     (p=0.271 n=10+10)
MVCCGet_Pebble/batch=true/versions=1/valueSize=8-10       2.54µs ± 2%    2.50µs ± 1%  -1.54%  (p=0.003 n=10+10)
MVCCGet_Pebble/batch=true/versions=10/valueSize=8-10      3.63µs ± 2%    3.38µs ± 2%  -6.99%  (p=0.000 n=10+10)
MVCCGet_Pebble/batch=true/versions=100/valueSize=8-10     7.90µs ± 1%    8.03µs ± 4%  +1.71%  (p=0.024 n=10+10)

name                                                    old speed      new speed      delta
MVCCGet_Pebble/batch=false/versions=1/valueSize=8-10    1.98MB/s ± 1%  1.97MB/s ± 0%  -0.51%  (p=0.024 n=9+7)
MVCCGet_Pebble/batch=false/versions=10/valueSize=8-10   1.63MB/s ± 1%  1.65MB/s ± 4%    ~     (p=0.137 n=10+10)
MVCCGet_Pebble/batch=false/versions=100/valueSize=8-10   807kB/s ± 2%   803kB/s ± 1%    ~     (p=0.290 n=10+10)
MVCCGet_Pebble/batch=true/versions=1/valueSize=8-10     3.15MB/s ± 2%  3.20MB/s ± 1%  +1.55%  (p=0.003 n=10+10)
MVCCGet_Pebble/batch=true/versions=10/valueSize=8-10    2.20MB/s ± 2%  2.36MB/s ± 1%  +7.38%  (p=0.000 n=10+9)
MVCCGet_Pebble/batch=true/versions=100/valueSize=8-10   1.01MB/s ± 1%  1.00MB/s ± 2%  -1.30%  (p=0.025 n=10+9)
```